Update AcceleoEvaluationVisitor.java
When last character of a protected content is a '}' it can be lost if a \r or \n character is before it.

With pattern "\\{(.)(.)?\\}" when protected content ends with a '}' after a \n or a \r, matcher captures the protected area marker '}' as 2nd character and appendReplacement overrides the '}' of the protected content. Using the pattern "\\{(n|r)(n)?\\}" solve this problem.